% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rent_model.R
\name{rent_pins}
\alias{rent_pins}
\title{Calcul du nombre de broches selon la règle de Rent}
\usage{
rent_pins(G, k, p)
}
\arguments{
\item{G}{nombre de blocs logiques}

\item{k}{constante de proportionnalité}

\item{p}{exposant de Rent}
}
\value{
nombre de broches externes
}
\description{
Calcul du nombre de broches selon la règle de Rent
}
