% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rent_fit.R
\name{rent_fit}
\alias{rent_fit}
\title{Estimation des paramètres de Rent (k et p)}
\usage{
rent_fit(G, P)
}
\arguments{
\item{G}{vecteur de blocs logiques}

\item{P}{vecteur de broches observées}
}
\value{
liste avec k et p estimés
}
\description{
Estimation des paramètres de Rent (k et p)
}
