% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rent_plot.R
\name{rent_plot}
\alias{rent_plot}
\title{Visualisation de la règle de Rent}
\usage{
rent_plot(G, P, fit)
}
\arguments{
\item{G}{vecteur de blocs logiques}

\item{P}{vecteur de broches observées}

\item{fit}{objet retourné par rent_fit}
}
\description{
Visualisation de la règle de Rent
}
